+++ sdk/data/azcosmos/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-* Added `ReadMany` API to read multiple items from a container in a single operation. See [PR TBD](https://github.com/Azure/azure-sdk-for-go/pull/TBD)
+* Added `ReadMany` API to read multiple items from a container in a single operation. See [PR 25516](https://github.com/Azure/azure-sdk-for-go/pull/25516)
 
 ### Breaking Changes